+++ src/recycle.c
@@ -160,17 +160,6 @@ GEN_DATA *new_gen_data(void)
     return gen;
 }
 
-void free_gen_data(GEN_DATA *gen)
-{
-    if (!IS_VALID(gen))
-	return;
-
-    INVALIDATE(gen);
-
-    gen->next = gen_data_free;
-    gen_data_free = gen;
-} 
-
 /* stuff for recycling extended descs */
 EXTRA_DESCR_DATA *extra_descr_free;